% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_events.R
\name{align_events}
\alias{align_events}
\title{Aligns data by events}
\usage{
align_events(
  df,
  events,
  dovars = names(df),
  leng_threshold,
  before,
  after,
  do_norm = FALSE,
  nbins = 6,
  normbin = 2
)
}
\arguments{
\item{df}{A data frame containing all data continuously along time, required columns: \code{"site", "date"}.}

\item{events}{A data frame with columns \code{idx_start} and \code{len}, specifying event start and length, referring to the row index of \code{df}.
\code{events} is the output of a function call to \code{get_consecutive}.}

\item{dovars}{A vector of character strings specifying which columns (by column name) of \code{df} to re-arrange.}

\item{leng_threshold}{An integer specifying the minum number of consecutive dates required to define an event.
All events of length lower than \code{leng_threshold} are dropped.}

\item{before}{An integer specifying the number of days before the event onset to be retained in re-arranged data}

\item{after}{An integer specifying the number of days after the event onset to be retained in re-arranged data}

\item{do_norm}{A logical specifying whether re-arranged data is to be normalised by the median value of the bin
(number of bins given by argument \code{nbins}) before the event onset, given by argument \code{normbin}. Defaults to \code{FALSE}.}

\item{nbins}{An integer, specifying the number of bins used to determine median values before event onset. Only used when code{do_norm=TRUE}. Defaults to 6.}

\item{normbin}{An integer, specifying the bin number just before the event onset, used for normalisation. Only used when code{do_norm=TRUE}. Defaults to 2.}
}
\value{
A named list of data frames (\code{list( "df_idx_event", "df_idx_event_aggbyidx_event")}) containing data from all events and \code{before + after}
dates (relative to event onset) with additional columns named \code{"inst"}, defining the event number (instance), and \code{"idx_event"}, defining
the date relative to the respective event onset. The data frame \code{"df_idx_event"} contains rearranged, but otherwise unchanged data (unless
\code{do_norm}=TRUE). The data frame \code{"df_idx_event_aggbyidx_event"} containes data aggregated across events with the mean and quantiles given for each
\code{"idx_event"}.
}
\description{
Uses a vectory specifying whether data falls into an event to reshape data, aligning by the onset of the event
}
