% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_modobs2.R
\name{analyse_modobs2}
\alias{analyse_modobs2}
\title{Analyse modelled values versus observed data.}
\usage{
analyse_modobs2(
  df,
  mod,
  obs,
  type = "points",
  relative = FALSE,
  shortsubtitle = FALSE,
  plot_subtitle = TRUE,
  plot_linmod = TRUE,
  pal = "viridis",
  ...
)
}
\arguments{
\item{df}{A data frame containing columns with names corresponding to arguments
\code{mod} and \code{obs}}

\item{mod}{A character string specifying the variable name (column) of the
modelled (simulated) values in data frame \code{df}.}

\item{obs}{A character string specifying the variable name (column) of the
observed values in data frame \code{df}.}

\item{type}{If \code{"points"}, uses \code{geom_points()}, if \code{"hex"}
uses \code{ggplot2::geom_hex()}, if \code{"heat"} uses adjusted
\code{geom_points()} with color indicating density, if \code{"density"} uses
\code{stat_density_2d()} to draw polygons of equal density.}

\item{relative}{A logical specifying whether the relative RMSE and bias (after
division by the mean) is to be showed in the subtitle labels.}

\item{shortsubtitle}{A logical specifying whether to display a reduced set of metrics
in the subtitle.}

\item{plot_subtitle}{A logical specifying whether to display any metrics. Defaults
to \code{TRUE}.}

\item{plot_linmod}{A logical specifying whether to display the fitted linear
regression as a red line. Defaults to \code{TRUE}.}

\item{pal}{A character string indicating the color palette. Currently available:
\code{"batlowW", "davos", "magma", "viridis", "cividis"}. Defaults to \code{"viridis"}.}

\item{nlabels}{An integer specifying how many points to be labelled, starting with points
that have the largest residuals from the linear regression fit. Only available
for \code{type == "points"}. Defaults to one.}
}
\description{
Calculates a set of performance statistics and optionally creates plots of modelled
versus observed values.
}
