% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_api.R
\name{configure}
\alias{configure}
\title{Configure Geco credentials, saving in keyring}
\usage{
configure(user, password, host)
}
\arguments{
\item{user}{Geco user name (email address). Defaults to GECO_API_USER environment variable}

\item{password}{Geco password (will prompt if not provided). Defaults to GECO_API_PASSWORD environment variable}

\item{host}{(optional) alternate host for API, only used for testing}
}
\description{
Configure Geco credentials, saving in keyring
}
