% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_api.R
\name{drop_configs}
\alias{drop_configs}
\title{Drop saved configurations
Warning! this will remove all saved configurations from the host.}
\usage{
drop_configs(user, host = "geco")
}
\arguments{
\item{user}{Optional username, provided as a string.}

\item{host}{Host identifier, defaults to 'geco'}
}
\description{
Drop saved configurations
Warning! this will remove all saved configurations from the host.
}
\seealso{
[list_configs()]
}
