% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_dataset.R
\name{extract_subsample_info}
\alias{extract_subsample_info}
\title{Characterize sampling information given dataset info, using standard column locations}
\usage{
extract_subsample_info(d)
}
\arguments{
\item{d}{data.frame containing result of `fetch_inference_dataset_info`}
}
\value{
data.frame with new columns containing information about the sample of data generated.
}
\description{
Characterize sampling information given dataset info, using standard column locations
}
