% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_biomarkers.R
\name{fetch_biomarkers}
\alias{fetch_biomarkers}
\title{Fetch biomarker data from the Generable API}
\usage{
fetch_biomarkers(
  project = NULL,
  project_version_id = NULL,
  measurement_name = NULL,
  annotate = T,
  annotate_doses = T
)
}
\arguments{
\item{project}{Project name}

\item{project_version_id}{Project version. If this is specified, the `project` argument is ignored.}

\item{measurement_name}{Vector of measurement names to return. `NULL` returns all measurements. Default
is `NULL`.}

\item{annotate}{if `TRUE`, annotate biomarker data with dose data. Default is `TRUE`.}

\item{annotate_doses}{if `TRUE`, annotate biomarker data with timing of dose administrations.
Default is `TRUE`.}
}
\value{
data.frame of biomarkers data for the project specified
}
\description{
Fetch biomarker data from the Generable API for a specific project.
}
\details{
This function retrieves biomarker data from the Generable API.
It requires authentication (see \code{\link{login}}) prior to use
and this pulls data from the Generable API.

A project can be specified by using the project name or a specific project version.
If a project is specified using the name, data is fetched for the latest version of the project.
If a project is specified using the project version, the project name is ignored if it
is also included as an argument.
}
