% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_dataset.R
\name{fetch_dataset}
\alias{fetch_dataset}
\title{Fetch the dataset from the Generable API for a model run}
\usage{
fetch_dataset(run_id, project = NULL, project_version_id = NULL)
}
\arguments{
\item{run_id}{Run id; required.}

\item{project}{Project name. If NULL, defaults to value of environment variable GECO_API_PROJECT}

\item{project_version_id}{Project version. If NULL, defaults to the most recent version of the project if provided, or the value of environment variable GECO_API_PROJECT_VERSION}
}
\value{
named list of `data.frame`s
}
\description{
Fetch the dataset from the Generable API for a model run.
}
\details{
A `dataset` is used by a model to generate a run. This function retrieves
the dataset used by a particular run. This returns the subjects, biomarkers,
and other information as a list of data.frames.

Authentication (see \code{\link{login}}) is required prior to using this function
and this pulls the metadata from the Generable API.
}
\note{
A project can be specified by using the project name or a specific project version.
\enumerate{
  \item If a project is specified using the name, data is fetched for the latest version of the project.
  \item If a project is specified using the project version, the project name is not required.
  \item If neither a project nor a project version is provided, the default project or project version is used. These are set by the environment variables GECO_API_PROJECT and GECO_API_PROJECT_VERSION
}
}
