% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_doses.R
\name{fetch_doses}
\alias{fetch_doses}
\title{Fetch dosing data for a Generable project}
\usage{
fetch_doses(project = NULL, project_version_id = NULL)
}
\arguments{
\item{project}{(chr) Name of project to return data for}

\item{project_version_id}{(chr) Optionally, a specific version of project data to return, if not the most recent}
}
\value{
data.frame of biomarkers data
}
\description{
Fetch dosing data for a Generable project
}
