% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_dataset.R
\name{fetch_inference_data}
\alias{fetch_inference_data}
\title{Fetch the datasets used for a particular model run.}
\usage{
fetch_inference_data(run_id, project = NULL, project_version_id = NULL)
}
\value{
a named list of data.frames, provided as input to the model fit.
}
\description{
Fetch the datasets used for a particular model run.
}
