% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_dataset.R
\name{fetch_inference_dataset_info}
\alias{fetch_inference_dataset_info}
\title{Fetch the model attributes for all datasets with at least one run in a project-version}
\usage{
fetch_inference_dataset_info(project = NULL, project_version_id = NULL)
}
\description{
Fetch the model attributes for all datasets with at least one run in a project-version
}
