% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_draws.R
\name{fetch_inference_draws}
\alias{fetch_inference_draws}
\title{Fetch draws of parameter or predictive quantities for a model run}
\usage{
fetch_inference_draws(
  parameter,
  project = NULL,
  project_version_id = NULL,
  run_id = NULL,
  predictive = F,
  type = c("posterior", "prior")
)
}
\description{
Fetch draws of parameter or predictive quantities for a model run
}
