% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_models.R
\name{fetch_inference_models}
\alias{fetch_inference_models}
\title{Fetch the model attributes for all models used at least once within a project-version}
\usage{
fetch_inference_models(project = NULL, project_version_id = NULL)
}
\description{
Fetch the model attributes for all models used at least once within a project-version
}
