% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_runs.R
\name{fetch_inference_runs}
\alias{fetch_inference_runs}
\title{Fetch the run attributes for all runs available for a project-version}
\usage{
fetch_inference_runs(project = NULL, project_version_id = NULL)
}
\description{
Fetch the run attributes for all runs available for a project-version
}
