% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_labs.R
\name{fetch_labs}
\alias{fetch_labs}
\title{Fetch labs data for a Generable project}
\usage{
fetch_labs(project = NULL, project_version_id = NULL, annotate = T)
}
\arguments{
\item{project}{(chr) Name of project to return data for}

\item{project_version_id}{(chr) Optionally, a specific version of project data to return, if not the most recent}

\item{annotate}{(bool) if TRUE, annotate returned biomarker data}
}
\value{
data.frame of labs data
}
\description{
Fetch labs data for a Generable project
}
