% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_biomarkers.R
\name{fetch_measurement_names}
\alias{fetch_measurement_names}
\title{Fetch distinct list of measurement names for a project}
\usage{
fetch_measurement_names(project = NULL, project_version_id = NULL)
}
\arguments{
\item{project}{(chr) Name of project to return data for}

\item{project_version_id}{(chr) Optionally, a specific version of project data to return, if not the most recent}
}
\value{
character vector of measurement names with at least one observation
}
\description{
Fetch distinct list of measurement names for a project
}
