% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkpd.R
\name{fetch_pkpd}
\alias{fetch_pkpd}
\title{Query Generable API for pkpd data in a standard format (biomarkers data and dosing data, merged)
returns a data.frame suitable for plotting and analysis.}
\usage{
fetch_pkpd(
  project = NULL,
  project_version_id = NULL,
  pd_measure = NULL,
  pk_measure = "concentration"
)
}
\arguments{
\item{project}{(chr) Name of project to return data for}

\item{project_version_id}{(chr) Optionally, a specific version of project data to return, if not the most recent}

\item{pd_measure}{measurement_name of PD measurement (defaults to NULL - no PD marker)}

\item{pk_measure}{measurement_name of PK measurement (defaults to 'conc', NULL indicates no PK marker)}
}
\value{
data.frame containing merged biomarker & dose data for the PK & PD parameter selected, with columns annotating cycles, time since last SDA, and measurement type.
}
\description{
Query Generable API for pkpd data in a standard format (biomarkers data and dosing data, merged)
returns a data.frame suitable for plotting and analysis.
}
