% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_projects.R
\name{fetch_projects}
\alias{fetch_projects}
\title{Lists available projects}
\usage{
fetch_projects()
}
\value{
data.frame containing information about each available project,
   one project per row.
}
\description{
Lists projects available to the authenticated user.
}
\details{
This function fetches the list of projects available to the
authenticated user from the Generable API. Included are the
project id, description, and other information pertient to
the project.
}
