% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_projectversions.R
\name{fetch_projectversions}
\alias{fetch_projectversions}
\title{Fetch information about versions available for a project}
\usage{
fetch_projectversions(project)
}
\arguments{
\item{project}{(chr) project id for which version information is requested}
}
\value{
data.frame with information about each project-version available for a project
}
\description{
Fetch information about versions available for a project
}
