% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_subjects.R
\name{fetch_subjects}
\alias{fetch_subjects}
\title{Fetch subject data from the Generable API}
\usage{
fetch_subjects(
  project = NULL,
  project_version_id = NULL,
  event_type = NULL,
  annotate = T,
  ...
)
}
\arguments{
\item{project}{Project name. If NULL, defaults to value of environment variable GECO_API_PROJECT}

\item{project_version_id}{Project version. If NULL, defaults to the most recent version of the project if provided, or the value of environment variable GECO_API_PROJECT_VERSION}

\item{event_type}{If this argument is provided, event data that matches this event type will be
joined into the return data.frame. Default is NULL (no event data).}

\item{annotate}{if `TRUE`, annotate subject data with dose data. Default is `TRUE`.}

\item{...}{Optional filters applied to subjects data, provided as name-value pairs where names are fields and values contain a subset of valid values
Example: trial_name = c('trial-A', 'trial-N'), performance_status = c(0,1)}
}
\value{
data.frame of subject-level data, including information about the trial and trial_arms
}
\description{
Fetch subject data from the Generable API for a specific project.
}
\details{
This function retrieves subject data from the Generable API. This includes
subject-level covariates.

As a convenience, event data can be included in the returned `data.frame` by
specifying the `event_type` argument; the data returned from \code{\link{fetch_events}}
matching the `event_type` will be merged into the returned data.

Authentication (see \code{\link{login}}) is reqiured prior to use
and this pulls subject data from the Generable API.
}
\note{
A project can be specified by using the project name or a specific project version.
\enumerate{
  \item If a project is specified using the name, data is fetched for the latest version of the project.
  \item If a project is specified using the project version, the project name is not required.
  \item If neither a project nor a project version is provided, the default project or project version is used. These are set by the environment variables GECO_API_PROJECT and GECO_API_PROJECT_VERSION
}
}
