% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_draws.R
\name{format_quantiles_as_widths}
\alias{format_quantiles_as_widths}
\title{Format a long summary of parameter quantiles (one record per run, parameter, and quantile) into a wide format (one record per run, parameter, and interval-width)}
\usage{
format_quantiles_as_widths(df)
}
\value{
a data.frame with new fields (compatible with ggdist plotting functions): .width, .lower, .upper, and .median
}
\description{
Format a long summary of parameter quantiles (one record per run, parameter, and quantile) into a wide format (one record per run, parameter, and interval-width)
}
