% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_era5_request.R
\name{gc_era5_request}
\alias{gc_era5_request}
\title{Download drivers from ECMWF}
\usage{
gc_era5_request(
  user,
  lat,
  lon,
  var,
  product,
  filename,
  start_date,
  end_date,
  method = "mean"
)
}
\arguments{
\item{user}{ECMWF Copernicus CDS user}

\item{lat}{latitude}

\item{lon}{longitude}

\item{var}{CDS variable of interest}

\item{product}{CDS product to query}

\item{filename}{output filename (CSV)}

\item{start_date}{start date as YYYY-MM-DD}

\item{end_date}{end date as YYYY-MM-DD}

\item{method}{aggregation method (mean, max, min, sum)}
}
\value{
A data frame with queried data including the date, position and
requested variable
}
\description{
Batch downloads ERA5 driver data from the ECMWF API (CDS Toolbox)
}
