% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consecutive.R
\name{get_consecutive}
\alias{get_consecutive}
\title{Identify events}
\usage{
get_consecutive(
  vec,
  merge_threshold = NA,
  leng_threshold = 3,
  do_merge = FALSE
)
}
\arguments{
\item{vec}{A vector of boolean values. Consecutive TRUE vakues designate an event.}

\item{merge_threshold}{An integer value specifying the threshold of the gap
length below in units of time steps which
gaps between events are ignored and the two events on either side of the gap
are merged into a single events. Defaults to NA (ignored). Is ignored if
\code{do_merge=FALSE}}

\item{leng_threshold}{An integer specifying the minimum length required for
creating an event. Defaults to 3.}

\item{do_merge}{A logical specifying whether to merge events if the gap between
them is small (smaller than \code{merge_threshold}).}
}
\value{
A data frame containing information about the start date and length
of each detected event
}
\description{
Identifies events as periods where of consecutively TRUE values in a boolean
vector.
}
