% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_biomarkers.R
\name{get_geco_biomarkers}
\alias{get_geco_biomarkers}
\title{Get biomarkers data for a Generable project}
\usage{
get_geco_biomarkers(
  project = NULL,
  project_version_id = NULL,
  measurement_name = NULL,
  annotate = T
)
}
\arguments{
\item{project}{(chr) Name of project to return data for}

\item{project_version_id}{(chr) Optionally, a specific version of project data to return, if not the most recent}

\item{measurement_name}{(chr, vector) Optionally, a list of measurement names to return}

\item{annotate}{(bool) if TRUE, annotate returned biomarker data}
}
\value{
data.frame of biomarkers data
}
\description{
Get biomarkers data for a Generable project
}
