% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_projectversions.R
\name{get_geco_projectversions}
\alias{get_geco_projectversions}
\title{Get information about versions available for a project}
\usage{
get_geco_projectversions(project)
}
\arguments{
\item{project}{(chr) project id for which version information is requested}
}
\value{
data.frame with information about each project-version available for a project
}
\description{
Get information about versions available for a project
}
