% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_subjects.R
\name{get_geco_subjects}
\alias{get_geco_subjects}
\title{Get subjects data for a Generable project}
\usage{
get_geco_subjects(project = NULL, project_version_id = NULL)
}
\arguments{
\item{project}{(chr) Name of project to return data for}

\item{project_version_id}{(chr) Optionally, a specific version of project data to return, if not the most recent}
}
\value{
data.frame of subject-level data, including information about the trial & trial_arms
}
\description{
Get subjects data for a Generable project
}
