% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_api.R
\name{list_configs}
\alias{list_configs}
\title{List saved configurations}
\usage{
list_configs(host = "geco")
}
\arguments{
\item{host}{Host identifier, defaults to 'geco'}
}
\description{
List saved configurations
}
