% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_runs.R
\name{list_parameter_names}
\alias{list_parameter_names}
\title{List the parameter names for a run}
\usage{
list_parameter_names(run_id, project = NULL, project_version_id = NULL)
}
\arguments{
\item{run_id}{Run id; required}

\item{project}{Project name. If NULL, defaults to value of environment variable GECO_API_PROJECT}

\item{project_version_id}{Project version. If NULL, defaults to the most recent version of the project if provided, or the value of environment variable GECO_API_PROJECT_VERSION}
}
\value{
vector of parameter names for the specified run
}
\description{
List the parameter names from the Generable API for a specific run.
}
\details{
A run is generated by a model and a dataset. This function retrieves the
names of all parameters for a specific run as a vector. This can be used in
\code{\link{fetch_quantiles}}.

Authentication (see \code{\link{login}}) is required prior to using this function
and this pulls the list of parameter names from the Generable API.

A project can be specified by using the project name or a specific project version.
\enumerate{
  \item If a project is specified using the name, data is fetched for the latest version of the project.
  \item If a project is specified using the project version, the project name is not required.
  \item If neither a project nor a project version is provided, the default project or project version is used. These are set by the environment variables GECO_API_PROJECT and GECO_API_PROJECT_VERSION
}
}
\seealso{
\code{\link{list_models}}, \code{\link{list_datasets}},
         \code{\link{fetch_quantiles}}, \code{\link{fetch_draws}}
}
