% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_runs.R
\name{list_predictive_names}
\alias{list_predictive_names}
\title{List the predictive names for a run}
\usage{
list_predictive_names(run_id, project = NULL, project_version_id = NULL)
}
\arguments{
\item{run_id}{Run id; required}

\item{project}{Project name}

\item{project_version_id}{Project version. If this is specified, the `project` argument is ignored.}
}
\value{
vector of predictive names for the specified run
}
\description{
List the predictive names from the Generable API for a specific run.
}
\details{
A run is generated by a model and a dataset. This function retrieves the
names of all predictive quantities for a specific run as a vector. This can be used in
\code{\link{fetch_quantiles}}.

Authentication (see \code{\link{login}}) is required prior to using this function
and this pulls the list of predictive names names from the Generable API.

A project can be specified by using the project name or a specific project version.
If a project is specified using the name, data is fetched for the latest version of the project.
If a project is specified using the project version, the project name is ignored if it
is also included as an argument.
}
\seealso{
\code{\link{list_models}}, \code{\link{list_datasets}},
         \code{\link{fetch_quantiles}}, \code{\link{fetch_draws}}
}
