% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_api.R
\name{login}
\alias{login}
\title{Login to the Generable API}
\usage{
login(user, password)
}
\arguments{
\item{user}{User email address. If not provided, will read the `GECO_API_USER` environment variable.}

\item{password}{User password. If not provided, will read the `GECO_API_PASSWORD` environment variable.}
}
\value{
OAuth 2.0 Bearer Token for the Generable API
}
\description{
Login to the Generable API
}
