% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map4.R
\name{plot_map4}
\alias{plot_map4}
\title{Plot a nice map with ggplot.}
\usage{
plot_map4(
  obj,
  varnam = NA,
  maxval = NA,
  breaks = NA,
  lonmin = -180,
  lonmax = 180,
  latmin = -90,
  latmax = 90,
  nbin = 10,
  legend_title = waiver(),
  legend_direction = "vertical",
  colorscale = viridis::viridis,
  color_ocean = "azure3",
  invert = -1,
  do_reproj = FALSE,
  hillshade = FALSE,
  rivers = FALSE,
  lakes = FALSE,
  coast = TRUE,
  ocean = FALSE,
  countries = FALSE,
  dir_ne = "~/data/naturalearth/",
  states = FALSE,
  scale = 110,
  make_discrete = TRUE,
  use_geom_raster = TRUE,
  is_boolean = FALSE,
  plot_title = waiver(),
  plot_subtitle = waiver(),
  combine = TRUE,
  ...
)
}
\arguments{
\item{obj}{An object, either a \code{SpatRaster} (returned from a \code{terra::rast()} function call),
or a list returned from a \code{rgeco::read_nc_onefile()} function call.}

\item{varnam}{A charachter string specifying the variable name. Optional and
used only if \code{obj} is a \code{SpatRaster} with multiple variables. If
\code{obj} is a rbeni-nc object (returned by \code{read_nc_onefile()}),
\code{varnam} must be provided (a character string specifying the variable
name in \code{obj$vars[[varnam]]}).}

\item{maxval}{A numeric value specifying the maximum value for which the color key is to be extended. Defaults
to \code{NA} (the 99\% quantile of values is used).}

\item{breaks}{A numeric vector specifying the breaks for the color scale. Defaults to \code{NA}, i.e. breaks
are determined automatically based on \code{nbin} and \code{maxval}.}

\item{lonmin}{Left edge (longitude, in degrees), defaults to -180.}

\item{lonmax}{Right edge (longitude, in degrees), defaults to 180.}

\item{latmin}{Lower edge (latitude, in degrees), defaults to -90.}

\item{latmax}{Upper edge (latitude, in degrees), defaults to 90.}

\item{nbin}{An integer specifying the number of bins used for the color key.}

\item{legend_title}{A character string specifying the legend title (annotation above the color key)}

\item{legend_direction}{Either \code{"vertical"} (default) or \code{"horizontal"}.}

\item{colorscale}{Either function that returns a set of colors or a vector of color names from which to interpolate.
Defaults to \code{virids::viridis}.}

\item{color_ocean}{A color specifyier for the fill color of the ocean layer. Defaults to \code{"azure"}.}

\item{invert}{One of 1 or -1, specifying the direction of the color scale. Defaults to -1.}

\item{do_reproj}{A boolean specifying whether to re-project the map to Robin projection}

\item{hillshade}{A logical specifying whether a hillshade layer should be added. Defaults to \code{FALSE}.}

\item{rivers}{A logical specifying whether to display rivers (the \code{ne_50m_rivers_lake_centerlines} layer from NaturalEarth.). Defaults to \code{FALSE}.}

\item{lakes}{A logical specifying whether to display rivers (the \code{ne_50m_lakes} layer from NaturalEarth). Defaults to \code{FALSE}.}

\item{coast}{A logical specifying whether to display coastlines (the \code{ne_50m_coastline} layer from NaturalEarth). Defaults to \code{TRUE}.}

\item{ocean}{A logical specifying whether to display the ocean layer from NaturalEarth). Defaults to \code{FALSE}.}

\item{countries}{A logical specifying whether to display country borders (the \code{ne_50m_admin_0_countries} layer from NaturalEarth). Defaults to \code{FALSE}.}

\item{dir_ne}{A character string specifying where to download Naturalearth layers. Once downloaded, they can be quickly loaded. Defaults to \code{"~/data/naturalearth/"}.}

\item{states}{A logical specifying whether to display sub-country administrative borders (e.g. US states) (the \code{ne_50m_admin_1_states_provinces} layer from NaturalEarth). Defaults to \code{FALSE}.}

\item{scale}{An integer specifying the resolutuion of NaturalEarth layers (coast, rivers, lakes). One of \code{110, 50, 10}. Defaults to \code{110} (coarsest resolution).
NaturalEarth layers for 110, 50, 10 m are used for low, medium, and high resolution (scale) layers, respectively. Defaults to \code{"small"}.}

\item{make_discrete}{A logical scpecifying whether data layer is to be made discrete for plotting with colors
of discrete bins. Defaults to \code{TRUE}.}

\item{use_geom_raster}{A logical specifying whether to use the function \code{geom_raster()} for plotting the raster layer.
Defaults to \code{TRUE}. If \code{FALSE}, \code{geom_tile()} is used. The latter can yield nicer results when data is sparse.}

\item{is_boolean}{A logical specifying whether the raster contains boolean values (either \code{TRUE} or \code{FALSE}). Defaults to \code{FALSE}.}

\item{plot_title}{A character string specifying the plot title}

\item{plot_subtitle}{A character string specifying the plot subtitle}

\item{combine}{A boolean specifying whether the map and the colorscale should be combined using cowplot.
Defaults to \code{TRUE}. If \code{FALSE}, a list of elements are retruned, where elements are the ggplot2 plot object
and the coloscale object returned by the call to \code{plot_discrete_cbar}.}

\item{...}{...}
}
\value{
A ggplot object for a global map plot.
}
\description{
Returns a cowplot object for a global map plot.
}
