% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_survival.R
\name{prep_km_data}
\alias{prep_km_data}
\title{Prepare kaplan-meier estimates in a tidy format}
\usage{
prep_km_data(data, formula)
}
\arguments{
\item{data}{dataframe in which to evaluate formula}

\item{formula}{formula to be passed to `survival::survfit`}
}
\value{
data.frame with kaplan-meier estimates over time
}
\description{
Prepare kaplan-meier estimates in a tidy format
}
