% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pkpd.R
\name{prep_pkpd_data}
\alias{prep_pkpd_data}
\title{Merge and annotate pkpd biomarkers data with dosing data
returns a data.frame suitable for plotting and analysis.}
\usage{
prep_pkpd_data(
  biomarkers_data,
  dose_data,
  pd_measure = NULL,
  pk_measure = NULL
)
}
\arguments{
\item{biomarkers_data}{data.frame containing biomarkers data}

\item{dose_data}{data.frame containing dose data}

\item{pd_measure}{measurement_name of PD measurement (defaults to NULL - no PD marker)}

\item{pk_measure}{measurement_name of PK measurement (defaults to 'conc', NULL indicates no PK marker)}
}
\value{
data.frame containing merged biomarker & dose data for the PK & PD parameter selected, with columns annotating cycles, time since last SDA, and measurement type.
}
\description{
Merge and annotate pkpd biomarkers data with dosing data
returns a data.frame suitable for plotting and analysis.
}
