% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{re-exports}
\alias{re-exports}
\alias{flog.logger}
\title{Re-exports}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{futile.logger}{\code{\link[futile.logger]{flog.logger}}}
}}

