% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrid_df.R
\name{regrid_df}
\alias{regrid_df}
\title{Tabular data regridding}
\usage{
regrid_df(df, res, lon_start, lon_end, lat_start, lat_end, varnam = NA)
}
\arguments{
\item{df}{A data frame. Must have columns named \code{'lon'} and \code{'lat'}.}

\item{res}{Resolution of the new grid in decimal degrees}

\item{lon_start}{The left (west) margin of the first (western-most) gridcell
along longitudes.}

\item{lon_end}{The right (east) margin of the first (eastern-most) gridcell
along longitudes.}

\item{lat_start}{The lower (south) margin of the first (southern-most) gridcell
along latitudes}

\item{lat_end}{The upper (north) margin of the first (northern-most) gridcell
along latitudes}

\item{varnam}{A character string specifying the variable name in the data frame.
When not specified, regrids all numeric variables.}
}
\value{
A data frame
}
\description{
For spatial data read into a data frame with longitude and latitude specified
for each observation, this function aggregates data to a new set of longitude/
latitude combinations, i.e., to a new grid. This effectively takes the mean
across all gridcells that have their mid-point within the coarse-resolution
gridcell and without area-weighing of constituent (fine-resolution) gridcells.
}
