% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeco-package.R
\docType{package}
\name{rgeco}
\alias{rgeco}
\title{rgeco: R package for accessing the Generable API}
\description{
The Generable API is a HTTP REST API that provides access
to data and inferences. The rgeco package provides convenient
access to the information in the API in an R-friendly format.
}
