% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_groups}
\alias{sample_groups}
\title{Filter the data to include only `n` groups}
\usage{
sample_groups(d, n, ..., replace = F)
}
\arguments{
\item{d}{data.frame from which to sample}

\item{n}{number of groups to sample}

\item{...}{unquoted variable names defining groups}

\item{replace}{boolean, whether to sample with replacement (default: F)}
}
\value{
subset of data with `n` groups selected
}
\description{
Filter the data to include only `n` groups
}
\examples{
 \dontrun{
 # plot data for a random individual, defined by `individual_id`
 d \%>\%
 sample_groups(n = 1, individual_id) \%>\%
 ggplot(., aes(...))
 }
}
