% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_biomarkers.R
\name{fetch_biomarkers}
\alias{fetch_biomarkers}
\title{Fetch biomarkers data for a Generable project}
\usage{
fetch_biomarkers(
  project = NULL,
  project_version_id = NULL,
  measurement_name = NULL,
  annotate = T,
  annotate_doses = T
)
}
\arguments{
\item{project}{(chr) Name of project to return data for}

\item{project_version_id}{(chr) Optionally, a specific version of project data to return, if not the most recent}

\item{measurement_name}{(chr, vector) Optionally, a list of measurement names to return. NULL returns all measurements}

\item{annotate}{(bool) if TRUE, annotate returned biomarker data}

\item{annotate_doses}{(bool) if TRUE, annotated returned biomarker data with timing of dose administrations, if available}
}
\value{
data.frame of biomarkers data
}
\description{
Fetch biomarkers data for a Generable project
}
