% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_events.R
\name{fetch_events}
\alias{fetch_events}
\title{Get event data for a Generable project}
\usage{
fetch_events(project = NULL, project_version_id = NULL, event_type = NULL)
}
\arguments{
\item{project}{(chr) Name of project to return data for}

\item{project_version_id}{(chr) Optionally, a specific version of project data to return, if not the most recent}

\item{event_type}{(chr) Optionally limit event_types to the names provided (example: "overall_survival"). NULL is unfiltered.}
}
\value{
data.frame with one record per subject and event type
}
\description{
Get event data for a Generable project
}
