% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_draws.R
\name{fetch_inference_quantiles}
\alias{fetch_inference_quantiles}
\title{Fetch quantiles of parameter or predictive quantity draws for a model run}
\usage{
fetch_inference_quantiles(
  parameter,
  project = NULL,
  project_version_id = NULL,
  run_id = NULL,
  predictive = F,
  type = c("posterior", "prior")
)
}
\description{
Fetch quantiles of parameter or predictive quantity draws for a model run
}
