% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_labs.R
\name{fetch_labs}
\alias{fetch_labs}
\title{Fetch lab data from the Generable API}
\usage{
fetch_labs(project = NULL, project_version_id = NULL, annotate = T)
}
\arguments{
\item{project}{Project name}

\item{project_version_id}{Project version. If this is specified, the `project` argument is ignored.}

\item{annotate}{if `TRUE`, annotate lab data with dose data. Default is `TRUE`.}
}
\value{
data.frame of lab data for the project specified
}
\description{
Fetch lab data from the Generable API for a specific project.
}
\details{
This function retrieves biomarker data from the Generable API.
It requires authentication (see \code{\link{login}}) prior to use
and this pulls data from the Generable API.

A project can be specified by using the project name or a specific project version.
If a project is specified using the name, data is fetched for the latest version of the project.
If a project is specified using the project version, the project name is ignored if it
is also included as an argument.
}
