% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_projects.R
\name{fetch_projects}
\alias{fetch_projects}
\title{Get listing of available GECO projects}
\usage{
fetch_projects()
}
\value{
data.frame of project-level data, including information about the project scope
}
\description{
Get listing of available GECO projects
}
