% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_projectversions.R
\name{fetch_projectversions}
\alias{fetch_projectversions}
\title{Lists project versions within a project}
\usage{
fetch_projectversions(project)
}
\value{
data.frame containing information about each available project version,
   one project version per row.
}
\description{
Lists project versions and provides information about each
project version
}
\details{
This function fetches the list of project versions within the specified
project from the Generable API. Included are the
project version id, description, and other information pertient to
the project version
}
