% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_runs.R
\name{find_runs}
\alias{find_runs}
\title{Find and filter runs on key features}
\usage{
find_runs(
  project = NULL,
  project_version_id = NULL,
  model_type = NULL,
  model_version = NULL,
  min_draws = 100,
  extra_fields = c()
)
}
\arguments{
\item{project}{Project name. If NULL, defaults to value of environment variable GECO_API_PROJECT}

\item{project_version_id}{Project version. If NULL, defaults to the most recent version of the project if provided, or the value of environment variable GECO_API_PROJECT_VERSION}

\item{model_type}{(character vector) filter to runs with this model type, as one of: joint, survival, biomarker. NULL to disable this filter.}

\item{model_version}{(character vector) filter to runs with this model version string. NULL to disable this filter.}

\item{min_draws}{(scalar int) filter to runs with >= this many draws combined across all chains. NULL to disable this filter.}

\item{extra_fields}{(character vector) names of additional fields to include in the summary. See results of \code{\link{list_datasets}}, \code{\link{list_runs}}, and \code{\link{list_models}} for available fields.}
}
\value{
a data frame with key metadata about the run.
}
\description{
This function is a higher-level wrapper around \code{\link{list_runs}} to aid in run discovery
}
\details{
A run is generated by a model and a dataset. This function retrieves the
key features for all runs meeting certain criteria within a project version.

The returned `data.frame` combines information about the run, model, and datasets
for each run.

Authentication (see \code{\link{login}}) is required prior to using this function
and this pulls the list of parameter names from the Generable API.

A project can be specified by using the project name or a specific project version.
\enumerate{
  \item If a project is specified using the name, data is fetched for the latest version of the project.
  \item If a project is specified using the project version, the project name is not required.
  \item If neither a project nor a project version is provided, the default project or project version is used. These are set by the environment variables GECO_API_PROJECT and GECO_API_PROJECT_VERSION
}
}
\seealso{
\code{\link{list_runs}}, \code{\link{list_models}}, \code{\link{list_datasets}},
         \code{\link{fetch_quantiles}}, \code{\link{fetch_draws}}
}
