% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_draws.R
\name{format_quantiles_as_widths}
\alias{format_quantiles_as_widths}
\title{Format a long summary of parameter quantiles (one record per run, parameter, and quantile) into a wide format (one record per run, parameter, and interval-width)}
\usage{
format_quantiles_as_widths(df)
}
\arguments{
\item{df}{a data.frame with quantile summary in long or denormalized format}
}
\value{
a data.frame with new fields (compatible with ggdist plotting functions): .width, .lower, .upper, and .value
}
\description{
Format a long summary of parameter quantiles (one record per run, parameter, and quantile) into a wide format (one record per run, parameter, and interval-width)
}
\note{
This function expects a data.frame in the format returned by \code{\link{fetch_quantiles}}.
}
