% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_runs.R
\name{list_runs}
\alias{list_runs}
\title{List runs from the Generable API}
\usage{
list_runs(project = NULL, project_version_id = NULL)
}
\arguments{
\item{project}{Project name}

\item{project_version_id}{Project version. If this is specified, the `project` argument is ignored.}
}
\value{
data.frame of run attributes for the project specified
}
\description{
List runs from the Generable API for a specific project.
}
\details{
A run is generated by a model and a dataset. This function retrieves the
attributes about all runs within a project version. The returned `data.frame`
contains information about what draws and quantiles are available and the model
and dataset ids.

The `run_draws` and `run_quantiles` columns contain two named lists of `parameter_names`
and `predictive_names`. The `parameter_names` is a list of the parameters of the model, e.g.
the sld parameters f, kg, and ks. The `predictive_names` is a list of the predicted quantities,
e.g. the predicted survival for each trial arm `predicted_survival_per_trial_arm`.

Authentication (see \code{\link{login}}) is required prior to using this function
and this pulls the metadata from the Generable API.

A project can be specified by using the project name or a specific project version.
If a project is specified using the name, data is fetched for the latest version of the project.
If a project is specified using the project version, the project name is ignored if it
is also included as an argument.
}
\seealso{
\code{\link{list_models}}, \code{\link{list_datasets}},
         \code{\link{fetch_quantiles}}, \code{\link{fetch_draws}}
}
