% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geco_api.R
\name{login}
\alias{login}
\title{Login to the Generable API}
\usage{
login(user, password)
}
\arguments{
\item{user}{(chr) user email}

\item{password}{(chr) user password}
}
\description{
Login to the Generable API
}
